# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/signup": {
            "post": {
                "description": "Sends an OTP code to the phoneNumber (fake OTP, just printed in the console)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Signup Method",
                "parameters": [
                    {
                        "description": "AuthSignupRequest model",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.AuthSignupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.AuthSignupResponse"
                        }
                    },
                    "400": {
                        "description": "Request body is not valid",
                        "schema": {
                            "$ref": "#/definitions/controllers_swagger.AuthSignupBadResponse"
                        }
                    }
                }
            }
        },
        "/auth/signup/confirm-otp": {
            "post": {
                "description": "Validate the PhoneNumber with the OtpCode alongside withit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Signup Confirm Otp Method",
                "parameters": [
                    {
                        "description": "AuthSignupConfirmOtpRequest model",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.AuthSignupConfirmOtpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.AuthTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Request body is not valid",
                        "schema": {
                            "$ref": "#/definitions/controllers_swagger.AuthSignupConfirmOtpBadResponse"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Current User Data using Jwt Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Current User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.GetCurrentUserResponse"
                        }
                    },
                    "400": {
                        "description": "Something Unknown Happend",
                        "schema": {
                            "$ref": "#/definitions/controllers_swagger.GetCurrentUserBadResponse"
                        }
                    }
                }
            }
        },
        "/user/{userId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get A User Data using Specified Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get User By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.GetCurrentUserResponse"
                        }
                    },
                    "400": {
                        "description": "Something Unknown Happend",
                        "schema": {
                            "$ref": "#/definitions/controllers_swagger.GetCurrentUserBadResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers_swagger.AuthSignupBadResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "string"
                },
                "succeed": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "controllers_swagger.AuthSignupConfirmOtpBadResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "string"
                },
                "succeed": {
                    "type": "boolean",
                    "example": false
                },
                "token": {
                    "type": "string",
                    "example": ""
                }
            }
        },
        "controllers_swagger.GetCurrentUserBadResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "string"
                },
                "succeed": {
                    "type": "boolean",
                    "example": false
                },
                "user": {
                    "type": "string",
                    "example": "null"
                }
            }
        },
        "dtos.AuthSignupConfirmOtpRequest": {
            "type": "object",
            "required": [
                "phoneNumber"
            ],
            "properties": {
                "otpCode": {
                    "type": "string",
                    "example": "1234"
                },
                "phoneNumber": {
                    "type": "string",
                    "maxLength": 11,
                    "minLength": 11,
                    "example": "09123456789"
                }
            }
        },
        "dtos.AuthSignupRequest": {
            "type": "object",
            "required": [
                "phoneNumber"
            ],
            "properties": {
                "phoneNumber": {
                    "type": "string",
                    "maxLength": 11,
                    "minLength": 11,
                    "example": "09123456789"
                }
            }
        },
        "dtos.AuthSignupResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "string"
                },
                "succeed": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dtos.AuthTokenResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Succeed"
                },
                "succeed": {
                    "type": "boolean",
                    "example": true
                },
                "token": {
                    "type": "string",
                    "example": "the token will be here"
                }
            }
        },
        "dtos.GetCurrentUserResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Succeed"
                },
                "succeed": {
                    "type": "boolean",
                    "example": true
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "createdDate": {
                    "type": "string",
                    "example": "2025-08-12T19:53:35.685Z"
                },
                "id": {
                    "type": "string",
                    "example": "689b9bbf5800ec55229e240b"
                },
                "phoneNumber": {
                    "type": "string",
                    "example": "09123456789"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}